#15 issue, splited item rarity tests into separate classes
